2007-03-26	Joel Sherrill <joel@OARcorp.com>
	* sp35/priinv.c: Remove warning.

+++ testsuites/sptests/sp35/priinv.c
@@ -82,7 +82,7 @@ const char *CallerName()
   #if defined(TEST_ON_RTEMS_45)
     TempName.u = *(uint32_t *)_Thread_Executing->Object.name;
   #else
-    TempName.u = _Thread_Executing->Object.name;
+    TempName.u = (uint32_t) _Thread_Executing->Object.name;
   #endif
   sprintf( buffer, "%c%c%c%c -- %d",
       TempName.c[0], TempName.c[1], TempName.c[2], TempName.c[3],
